Moved set_default_options to coredata
